feat: Add PostContext to fetch and provide posts data
This commit adds a new file `PostContext.jsx` in the `client/src/contexts` directory. The `PostContext` component is created using the `createContext` function from React. It also includes a `PostContextProvider` component that wraps the children components and fetches the posts data from the server using Axios. The fetched posts data is stored in the `posts` state using the `useState` hook.

In the `client/src/main.jsx` file, the `PostContextProvider` component is imported and wrapped around the `BrowserRouter` component to provide the posts data to the application.

In the `client/src/pages/Home.jsx` file, the `PostContext` is imported and the `posts` data is accessed using the `useContext` hook. The fetched posts data is then rendered in the component.

In the `server/models/PostModel.js` file, the `author` field is updated to use `Schema.Types.ObjectId` and reference the `User` model.

In the `server/routes/postsRouter.js` file, the `isAuth` middleware is removed from the `getAllPosts` and `getPostById` routes.

This commit addresses the need to fetch and provide posts data to the application using the `PostContext` component.