feat: Add PostInfo component and update PostModel
This commit adds the `PostInfo` component to the `client/src/components` directory. The component renders basic information about a post.

In the `server/models/PostModel.js` file, the `author` field is updated to use `Schema.Types.ObjectId` and reference the `User` model.

These changes address the need to display post information and update the post model to use the correct data type for the author field.